Comments / improved use of Response from server.
Added significant comments to the DJ service. Additionally, the server
makes better uses of the response object when replying to the DJ. See
comments in code for specifics, but generally the following can be
expected. Error is set, Message is set if an error occurs. Result stores
how many entries are coming back if applicable. Result stores how many
songs were added/removed successfully when applicable.